Add a log function to the repository, change GitCommit to parse that, and add in infrastructure for tests.
The GitCommit class now parses a list of potential commit descriptions, rather than a single one.  That mechanism is now used by GitCommit, but the usage there is as-yet untested, and likely broken.  The basic infrastructure has been put in place for writing tests, including ivy files to download the junit jar, library changes to include Gosu and junit at runtime as well as the src and test-src directories, and a skeletal GosuGitSuite class with one initial test.